Alteracao nas cores
