+++ CalculatorIOS/CalculatorIOS/SceneDelegate.swift
@@ -15,8 +15,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
-        let mainViewController = ViewController()
-        mainViewController.view.backgroundColor = .systemBlue
+        let mainViewController = CalculatorViewController()
+        mainViewController.view.backgroundColor = .black
         
         let navController = UINavigationController(rootViewController: mainViewController)
         window.rootViewController = navController
